Merged LORA Code into Scanner Code
Added the LORA sending code to the existing code, could not achieve a stable connection to the TTN jet, issues are more likely Gateway related, since with the example sketch for sending "Hello World" it was aswell not possible to achieve a stable connection! 

Also organised the code to make it more readable.